FEAT: deletion exception

+++ demo/src/main/java/com/example/demo/service/UserService.java
@@ -2,8 +2,10 @@ package com.example.demo.service;
 
 import com.example.demo.entities.User;
 import com.example.demo.repositories.UserRepository;
+import com.example.demo.service.exceptions.DatabaseException;
 import com.example.demo.service.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,7 +30,12 @@ public class UserService {
     }
 
     public void delete(Long id){
-        userRepository.deleteById(id);
+      try {
+          if (!userRepository.existsById(id)) throw new ResourceNotFoundException(id);
+          userRepository.deleteById(id);
+      } catch (DataIntegrityViolationException e){
+          throw new DatabaseException(e.getMessage());
+      }
     }
 
     public User update(Long id, User user){
